Move LegacyTagHelperCompletionService out of MS.CA.Razor.Workspaces
