+++ src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/Serialization/CompletionListSerializationBenchmark.cs
@@ -26,7 +26,7 @@ public class CompletionListSerializationBenchmark
     public CompletionListSerializationBenchmark()
     {
         var tagHelperFactsService = new TagHelperFactsService();
-        var completionService = new LanguageServerTagHelperCompletionService(tagHelperFactsService);
+        var completionService = new LspTagHelperCompletionService(tagHelperFactsService);
         var htmlFactsService = new DefaultHtmlFactsService();
         var optionsMonitor = new BenchmarkOptionsMonitor<RazorLSPOptions>(RazorLSPOptions.Default);
         var tagHelperCompletionProvider = new TagHelperCompletionProvider(completionService, htmlFactsService, tagHelperFactsService, optionsMonitor);
