+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Completion/LanguageServerTagHelperCompletionServiceTest.cs
@@ -10,13 +10,15 @@ using System.Linq;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.Components;
 using Microsoft.AspNetCore.Razor.Test.Common;
+using Microsoft.CodeAnalysis.Razor.Completion;
+using Microsoft.VisualStudio.Editor.Razor;
 using Xunit;
 using Xunit.Abstractions;
 using static Microsoft.AspNetCore.Razor.Language.CommonMetadata;
 
-namespace Microsoft.VisualStudio.Editor.Razor;
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion;
 
-public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) : ToolingTestBase(testOutput)
+public class LanguageServerTagHelperCompletionServiceTest(ITestOutputHelper testOutput) : ToolingTestBase(testOutput)
 {
     [Fact]
     [WorkItem("https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1452432")]
@@ -653,10 +655,7 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
                 .TagOutputHint("table")
                 .Build(),
         };
-        var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>()
-        {
-            ["table"] = new HashSet<TagHelperDescriptor>(),
-        });
+        var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>());
 
         var existingCompletions = new[] { "table" };
         var completionContext = BuildElementCompletionContext(
@@ -728,11 +727,10 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         };
         var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>()
         {
-            ["my-table"] = new HashSet<TagHelperDescriptor> { documentDescriptors[0] },
-            ["table"] = new HashSet<TagHelperDescriptor>(),
+            ["my-table"] = new HashSet<TagHelperDescriptor> { documentDescriptors[0] }
         });
 
-        var existingCompletions = new[] { "table" };
+        var existingCompletions = new[] { "table", "div" };
         var completionContext = BuildElementCompletionContext(
             documentDescriptors,
             existingCompletions,
@@ -763,7 +761,6 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>()
         {
             ["th:li"] = new HashSet<TagHelperDescriptor> { documentDescriptors[1], documentDescriptors[0] },
-            ["li"] = new HashSet<TagHelperDescriptor>(),
         });
 
         var existingCompletions = new[] { "li" };
@@ -798,7 +795,6 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         {
             ["th:superli"] = new HashSet<TagHelperDescriptor> { documentDescriptors[0] },
             ["th:li"] = new HashSet<TagHelperDescriptor> { documentDescriptors[1] },
-            ["li"] = new HashSet<TagHelperDescriptor>(),
         });
 
         var existingCompletions = new[] { "li" };
@@ -835,7 +831,6 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         {
             ["myli"] = new HashSet<TagHelperDescriptor> { documentDescriptors[0] },
             ["MYLI"] = new HashSet<TagHelperDescriptor> { documentDescriptors[1] },
-            ["li"] = new HashSet<TagHelperDescriptor> { },
         });
 
         var existingCompletions = new[] { "li" };
@@ -905,7 +900,6 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>()
         {
             ["superli"] = new HashSet<TagHelperDescriptor>() { documentDescriptors[0], documentDescriptors[1] },
-            ["li"] = new HashSet<TagHelperDescriptor>(),
         });
 
         var existingCompletions = new[] { "li" };
@@ -938,7 +932,6 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>()
         {
             ["th:superli"] = new HashSet<TagHelperDescriptor>() { documentDescriptors[0], documentDescriptors[1] },
-            ["th:li"] = new HashSet<TagHelperDescriptor>() { documentDescriptors[1] },
         });
 
         var existingCompletions = new[] { "th:li" };
@@ -1174,8 +1167,8 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         var completionContext = BuildElementCompletionContext(
             documentDescriptors,
             existingCompletions: Enumerable.Empty<string>(),
-            containingTagName: "parent-tag",
-            containingParentTagName: null);
+            containingTagName: "child-tag",
+            containingParentTagName: "parent-tag");
         var service = CreateTagHelperCompletionFactsService();
 
         // Act
@@ -1236,7 +1229,7 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         });
 
         var existingCompletions = new[] { "p", "em" };
-        var completionContext = BuildElementCompletionContext(documentDescriptors, existingCompletions, containingTagName: "div", containingParentTagName: "thing");
+        var completionContext = BuildElementCompletionContext(documentDescriptors, existingCompletions, containingTagName: "thing", containingParentTagName: "div");
         var service = CreateTagHelperCompletionFactsService();
 
         // Act
@@ -1264,7 +1257,7 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
             ["bold"] = new HashSet<TagHelperDescriptor>(),
         });
 
-        var completionContext = BuildElementCompletionContext(documentDescriptors, Enumerable.Empty<string>(), containingTagName: "div", containingParentTagName: "");
+        var completionContext = BuildElementCompletionContext(documentDescriptors, Enumerable.Empty<string>(), containingTagName: "", containingParentTagName: "div");
         var service = CreateTagHelperCompletionFactsService();
 
         // Act
@@ -1294,7 +1287,7 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
             ["div"] = new HashSet<TagHelperDescriptor> { documentDescriptors[0] }
         });
 
-        var completionContext = BuildElementCompletionContext(documentDescriptors, Enumerable.Empty<string>(), containingTagName: "div", containingParentTagName: "");
+        var completionContext = BuildElementCompletionContext(documentDescriptors, Enumerable.Empty<string>(), containingTagName: "", containingParentTagName: "div");
         var service = CreateTagHelperCompletionFactsService();
 
         // Act
@@ -1330,7 +1323,7 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
             ["div"] = new HashSet<TagHelperDescriptor> { documentDescriptors[0], documentDescriptors[1] },
         });
 
-        var completionContext = BuildElementCompletionContext(documentDescriptors, Enumerable.Empty<string>(), containingTagName: "div", containingParentTagName: "");
+        var completionContext = BuildElementCompletionContext(documentDescriptors, Enumerable.Empty<string>(), containingTagName: "", containingParentTagName: "div");
         var service = CreateTagHelperCompletionFactsService();
 
         // Act
@@ -1340,10 +1333,103 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         AssertCompletionsAreEquivalent(expectedCompletions, completions);
     }
 
-    private static LegacyTagHelperCompletionService CreateTagHelperCompletionFactsService()
+    [Fact]
+    public void GetElementCompletions_MustSatisfyAttributeRules_WithAttributes()
+    {
+        // Arrange
+        var documentDescriptors = new[]
+        {
+            TagHelperDescriptorBuilder.Create("FormTagHelper", "TestAssembly")
+                .TagMatchingRuleDescriptor(rule => rule
+                    .RequireTagName("form")
+                .RequireAttributeDescriptor(builder =>
+                    {
+                        builder.Name = "asp-route-";
+                        builder.NameComparisonMode = RequiredAttributeDescriptor.NameComparisonMode.PrefixMatch;
+                    }))
+                .Build(),
+        };
+        var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>()
+        {
+            ["form"] = new HashSet<TagHelperDescriptor> { documentDescriptors[0] }
+        });
+
+        var attributes = ImmutableArray.Create(
+            KeyValuePair.Create("asp-route-id", "123"));
+
+        var completionContext = BuildElementCompletionContext(documentDescriptors, existingCompletions: ["form"], containingTagName: "", containingParentTagName: "div", attributes: attributes);
+        var service = CreateTagHelperCompletionFactsService();
+
+        // Act
+        var completions = service.GetElementCompletions(completionContext);
+
+        // Assert
+        AssertCompletionsAreEquivalent(expectedCompletions, completions);
+    }
+
+    [Fact]
+    public void GetElementCompletions_MustSatisfyAttributeRules_NoAttributes()
+    {
+        // Arrange
+        var documentDescriptors = new[]
+        {
+            TagHelperDescriptorBuilder.Create("FormTagHelper", "TestAssembly")
+                .TagMatchingRuleDescriptor(rule => rule
+                    .RequireTagName("form")
+                .RequireAttributeDescriptor(builder =>
+                    {
+                        builder.Name = "asp-route-";
+                        builder.NameComparisonMode = RequiredAttributeDescriptor.NameComparisonMode.PrefixMatch;
+                    }))
+                .Build(),
+        };
+        var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>());
+
+        var completionContext = BuildElementCompletionContext(documentDescriptors, existingCompletions: ["form"], containingTagName: "", containingParentTagName: "div");
+        var service = CreateTagHelperCompletionFactsService();
+
+        // Act
+        var completions = service.GetElementCompletions(completionContext);
+
+        // Assert
+        AssertCompletionsAreEquivalent(expectedCompletions, completions);
+    }
+
+    [Fact]
+    public void GetElementCompletions_MustSatisfyAttributeRules_NoAttributes_AllowedIfNotHtml()
+    {
+        // Arrange
+        var documentDescriptors = new[]
+        {
+            TagHelperDescriptorBuilder.Create("ComponentTagHelper", "TestAssembly")
+                .TagMatchingRuleDescriptor(rule => rule
+                    .RequireTagName("component")
+                .RequireAttributeDescriptor(builder =>
+                    {
+                        builder.Name = "type";
+                        builder.NameComparisonMode = RequiredAttributeDescriptor.NameComparisonMode.PrefixMatch;
+                    }))
+                .Build(),
+        };
+        var expectedCompletions = ElementCompletionResult.Create(new Dictionary<string, HashSet<TagHelperDescriptor>>()
+        {
+            ["component"] = new HashSet<TagHelperDescriptor>() { documentDescriptors[0] },
+        });
+
+        var completionContext = BuildElementCompletionContext(documentDescriptors, Enumerable.Empty<string>(), containingTagName: "", containingParentTagName: "div");
+        var service = CreateTagHelperCompletionFactsService();
+
+        // Act
+        var completions = service.GetElementCompletions(completionContext);
+
+        // Assert
+        AssertCompletionsAreEquivalent(expectedCompletions, completions);
+    }
+
+    private static LspTagHelperCompletionService CreateTagHelperCompletionFactsService()
     {
         var tagHelperFactsService = new TagHelperFactsService();
-        var completionFactService = new LegacyTagHelperCompletionService(tagHelperFactsService);
+        var completionFactService = new LspTagHelperCompletionService(tagHelperFactsService);
 
         return completionFactService;
     }
@@ -1378,17 +1464,20 @@ public class LegacyTagHelperCompletionServiceTest(ITestOutputHelper testOutput) 
         string containingTagName,
         string containingParentTagName = "body",
         bool containingParentIsTagHelper = false,
-        string tagHelperPrefix = "")
+        string tagHelperPrefix = "",
+        ImmutableArray<KeyValuePair<string, string>> attributes = default)
     {
+        attributes = attributes.NullToEmpty();
+
         var documentContext = TagHelperDocumentContext.Create(tagHelperPrefix, descriptors);
         var completionContext = new ElementCompletionContext(
             documentContext,
             existingCompletions,
             containingTagName,
-            attributes: ImmutableArray<KeyValuePair<string, string>>.Empty,
+            attributes,
             containingParentTagName: containingParentTagName,
             containingParentIsTagHelper: containingParentIsTagHelper,
-            inHTMLSchema: (tag) => tag == "strong" || tag == "b" || tag == "bold" || tag == "li" || tag == "div");
+            inHTMLSchema: (tag) => tag == "strong" || tag == "b" || tag == "bold" || tag == "li" || tag == "div" || tag == "form");
 
         return completionContext;
     }
