+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Completion/LegacyRazorCompletionEndpointTest.cs
@@ -34,7 +34,7 @@ public class LegacyRazorCompletionEndpointTest : LanguageServerTestBase
     {
         // Working around strong naming restriction.
         var tagHelperFactsService = new TagHelperFactsService();
-        var tagHelperCompletionService = new LanguageServerTagHelperCompletionService(tagHelperFactsService);
+        var tagHelperCompletionService = new LspTagHelperCompletionService(tagHelperFactsService);
 
         var completionProviders = new IRazorCompletionItemProvider[]
         {
